splitting talking head into separate files in order to allow using the service inside of other threejs experiences such as games - a later commit would move the initialization of threejs itself to a separate file
